requirements.txt
flask
flask-cors
python-dotenv
requests
pymupdf
together
openai